:hammer: change component Sidebar to Header and Header to Title
