+++ src/components/Title.jsx
@@ -1,0 +1,11 @@
+import "../styles/components/title.sass";
+
+export function Title(){
+   return (
+      <div className="title">
+         <h2>Hi, my name is</h2>
+         <h1>Jessica Adamek</h1>
+         <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Saepe odio quas laborum? Eos odit officia tempora natus quae, ducimus</p>
+      </div>
+   )
+}
